More basic stuff
Added funds property to PlayerrConfiguration, added some stuff to the
GameplayController to make this more manageable.

What is missing now is the turn-based game engine and coloring the
MainMap. Not sure how to proceed here since the solution I can think of
seems like it would just overwrite the fxml.

I think we have the necessary components, now it's just putting them
together.